refactor(core,provider,tauri): 清理代码并修复 Clippy 警告
- 修复类型转换中的精度损失警告，添加适当的 allow 属性
- 使用更简洁的语法（如 if let Some(x) = ... 替代 match）
- 移除未使用的代码和死代码属性
- 优化字符串构建，使用 write! 宏避免临时分配
- 更新依赖版本（url 2.5.8，tauri 插件等）
- 统一错误处理模式，移除冗余的网络错误快捷方法
- 简化模式匹配，合并相同分支
- 添加必要的文档注释和 panic 说明
- 修复导入导出服务中的凭证加载逻辑
- 优化 DNSSEC 检查的状态判断逻辑
- 清理迁移模板文件中的占位代码

+++ dns-orchestrator-core/src/services/toolbox/dns_propagation.rs
@@ -134,11 +134,16 @@ fn calculate_consistency(results: &[DnsPropagationServerResult]) -> (f32, Vec<St
 
     let total = successful_results.len();
     let max_count = value_counts.values().max().copied().unwrap_or(0);
-    let consistency = (max_count as f32 / total as f32) * 100.0;
+    // usize -> f64: these are small counts of DNS server results (typically < 20),
+    // well within f64's precise integer range (up to 2^52)
+    #[allow(clippy::cast_precision_loss)]
+    let consistency = (max_count as f64 / total as f64) * 100.0;
 
     let unique_values: Vec<_> = value_counts.keys().cloned().collect();
 
-    (consistency, unique_values)
+    #[allow(clippy::cast_possible_truncation)]
+    // f64 -> f32: consistency is a percentage (0.0..=100.0), well within f32 range
+    (consistency as f32, unique_values)
 }
 
 /// DNS 传播检查
@@ -162,6 +167,8 @@ pub async fn dns_propagation_check(
                     dns_lookup(&domain, &record_type, Some(&server.ip)),
                 )
                 .await;
+                // u128 -> u64: elapsed millis for a DNS query will never exceed u64::MAX
+                #[allow(clippy::cast_possible_truncation)]
                 let elapsed = query_start.elapsed().as_millis() as u64;
 
                 match result {
@@ -196,6 +203,8 @@ pub async fn dns_propagation_check(
     // 计算一致性
     let (consistency_percentage, unique_values) = calculate_consistency(&results);
 
+    // u128 -> u64: elapsed millis for DNS propagation check will never exceed u64::MAX
+    #[allow(clippy::cast_possible_truncation)]
     let total_time_ms = start_time.elapsed().as_millis() as u64;
 
     Ok(DnsPropagationResult {
